requirements.txt
langchain
langchain-ollama
langchain-mcp-adapters
langgraph
fastmcp
gspread
google-auth
pyyaml